refactor: Extract login form validation schema into a separate file
